+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Fixed parsing of emoji shortcodes.
 - Re-sign fetcher request after redirection.
+- Don't replace emojis inside words.
 
 ## [2.16.0] - 2024-04-16